Update all services schema tests
