+++ tests/schemas/test_network_service_invalid_schema_cases.py
@@ -1,0 +1,46 @@
+from typing import Literal
+
+from pytest_cases import case, parametrize
+
+from fed_reg.provider.schemas_extended import (
+    NetworkCreateExtended,
+    NetworkQuotaCreateExtended,
+)
+from fed_reg.service.enum import ServiceType
+from tests.utils import random_lower_string
+
+
+class CaseInvalidAttr:
+    @case(tags=["base"])
+    @parametrize(attr=("endpoint", "name"))
+    def case_none(self, attr: str) -> tuple[str, None]:
+        return attr, None
+
+    @case(tags=["base"])
+    def case_endpoint(self) -> tuple[Literal["endpoint"], None]:
+        return "endpoint", random_lower_string()
+
+    @case(tags=["base"])
+    @parametrize(value=(i for i in ServiceType if i != ServiceType.NETWORK))
+    def case_type(self, value: ServiceType) -> tuple[Literal["type"], ServiceType]:
+        return "type", value
+
+    @case(tags=["create_extended"])
+    def case_dup_quotas(
+        self, network_quota_create_ext_schema: NetworkQuotaCreateExtended
+    ) -> tuple[Literal["quotas"], list[NetworkQuotaCreateExtended], str]:
+        return (
+            "quotas",
+            [network_quota_create_ext_schema, network_quota_create_ext_schema],
+            "Multiple quotas on same project",
+        )
+
+    @case(tags=["create_extended"])
+    def case_dup_networks(
+        self, network_create_ext_schema: NetworkCreateExtended
+    ) -> tuple[Literal["networks"], list[NetworkCreateExtended], str]:
+        return (
+            "networks",
+            [network_create_ext_schema, network_create_ext_schema],
+            "There are multiple items with identical uuid",
+        )
